TISTUD-5060 TISTUD-5101 Fix two bugs found during my demo

+++ plugins/com.aptana.git.core/src/com/aptana/git/core/github/IGithubPullRequest.java
@@ -18,7 +18,7 @@ public interface IGithubPullRequest
 
 	public URL getURL() throws MalformedURLException;
 
-	public int getNumber();
+	public Long getNumber();
 
 	public String getTitle();
